requirements.txt
camel-ai[all]==0.2.38
paddleocr == 3.0.0
wikipedia == 1.4.0
